+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to the "lua-tags" extension will be documented in this file.
 1. fix signature help parameter index incorrect
 2. fix workspace symbol send too many symbol infomation
 3. add sub function declaration support
+4. filter other document local symbol at definition、auto completion、hover
 
 ## [1.0.0] 2019-12-12